extract HomeEnrollmentsDashboard into its own file
EnrollmentDisplay.tsx had grown to ~836 lines because it housed both
the home dashboard and the program dashboard. Pull the home half out:
HomeEnrollmentsDashboard.tsx now owns its component, EnrollmentExpandCollapse,
the home-only styled components, and SUPPORT_EMAIL. EnrollmentDisplay.tsx
is now 836 → 548 lines and dispatches to either HomeEnrollmentsDashboard
or ProgramEnrollmentDisplay.

Phase 4 and Phase 5 will extract the program and contract dashboards into
sibling files, leaving EnrollmentDisplay.tsx as a thin router (or removable
entirely once each page can import its dashboard directly).

DashboardCardStyled is duplicated across the two files for now — small
styled wrapper, scheduled to be replaced by CoursewareCard in Phase 7.
